# static/swagger/v1/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SkiResort",
    "description": "API for Web Course Project",
    "version": "v1"
  },
  "paths": {
...
    "/lifts": {
      "get": {
        "tags": [
          "lifts"
        ],
        "summary": "Get information about all lifts",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Information about all lifts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lift"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "lifts"
        ],
        "summary": "Add a new lift",
        "description": "Note that the names of lifts should be unique.",
        "parameters": [
          {
            "name": "liftName",
            "in": "query",
            "description": "Name of the new lift",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isOpen",
            "in": "query",
            "description": "Is the new lift currently working or not",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "seatsAmount",
            "in": "query",
            "description": "The amount of seats in the new lift",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "liftingTime",
            "in": "query",
            "description": "The time the new lift needs to lift from the beginning to the end",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "The added lift with assigned ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lift"
                }
              }
            }
          },
          "400": {
            "description": "A lift with such name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/lifts/{liftName}": {
      "get": {
        "tags": [
          "lifts"
        ],
        "summary": "Get information about a lift by it's name",
        "parameters": [
          {
            "name": "liftName",
            "in": "path",
            "description": "Name of the lift",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Lift with specified name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiftWithSlopes"
                }
              }
            }
          },
          "404": {
            "description": "Lift with specified name not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "lifts"
        ],
        "summary": "Update information about an existing lift",
        "parameters": [
          {
            "name": "liftName",
            "in": "path",
            "description": "Name of the lift to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IsOpen",
            "in": "query",
            "description": "Is the lift working right now or not",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "SeatsAmount",
            "in": "query",
            "description": "The amount of seats in the lift",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "LiftingTime",
            "in": "query",
            "description": "The time lift needs to lift from the beginning to the end",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "The lift was successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "404": {
            "description": "A lift with specified name was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "lifts"
        ],
        "summary": "Delete a lift by it's name",
        "parameters": [
          {
            "name": "liftName",
            "in": "path",
            "description": "Name of the lift",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Lift was successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "404": {
            "description": "Lift with specified name not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "lifts"
        ],
        "summary": "Add a connected slope to the lift",
        "parameters": [
          {
            "name": "liftName",
            "in": "path",
            "description": "Name of the lift to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slopeName",
            "in": "query",
            "description": "Name of the slope to add",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "The lift was successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IActionResult"
                }
              }
            }
          },
          "404": {
            "description": "A lift or a slope with specified name was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IActionResult": {
        "type": "object",
        "additionalProperties": false
      },
...
      "Lift": {
        "type": "object",
        "properties": {
          "liftID": {
            "type": "integer",
            "description": "Lift ID",
            "format": "int32"
          },
          "liftName": {
            "type": "string",
            "description": "Lift name",
            "nullable": true
          },
          "isOpen": {
            "type": "boolean",
            "description": "Is the lift working right now or not"
          },
          "seatsAmount": {
            "type": "integer",
            "description": "The amount of seats in the lift",
            "format": "int32"
          },
          "liftingTime": {
            "type": "integer",
            "description": "The time lift needs to lift from the beginning to the end",
            "format": "int32"
          },
          "queueTime": {
            "type": "integer",
            "description": "Current time in queue to the lift",
            "format": "int32",
            "default": 0
          }
        },
        "additionalProperties": false,
        "description": "Information about a lift"
      },
      "LiftWithSlopes": {
        "type": "object",
        "properties": {
          "liftID": {
            "type": "integer",
            "description": "Lift ID",
            "format": "int32"
          },
          "liftName": {
            "type": "string",
            "description": "Lift name",
            "nullable": true
          },
          "isOpen": {
            "type": "boolean",
            "description": "Is the lift working right now or not"
          },
          "seatsAmount": {
            "type": "integer",
            "description": "The amount of seats in the lift",
            "format": "int32"
          },
          "liftingTime": {
            "type": "integer",
            "description": "The time lift needs to lift from the beginning to the end",
            "format": "int32"
          },
          "queueTime": {
            "type": "integer",
            "description": "Current time in queue to the lift",
            "format": "int32",
            "default": 0
          },
          "connectedSlopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Slope"
            },
            "description": "Slopes connected to the lift",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Information about a lift"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "Slope": {
        "type": "object",
        "properties": {
          "slopeID": {
            "type": "integer",
            "description": "Slope ID",
            "format": "int32"
          },
          "slopeName": {
            "type": "string",
            "description": "Slope name",
            "nullable": true
          },
          "isOpen": {
            "type": "boolean",
            "description": "Is the slope working right now or not"
          },
          "difficultyLevel": {
            "type": "integer",
            "description": "The difficulty level of the slope",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Information about a slope"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 1safsfsdfdfd\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}